pyproject.toml
[tool.poetry]
name = "situational-task-5"
version = "0.1.0"
description = ""
authors = ["Bolot Zhusupekov <bolot.zhusupekov07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-cacheops = "^7.0.2"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
psycopg2 = "^2.9.9"
django-allauth = "^0.63.3"
isort = "^5.13.2"
djangorestframework-simplejwt = "^5.3.1"
cryptography = "^42.0.7"
requests = "^2.32.3"
openapi = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
